发送 退出xx 消息，优化推送流程异步化。

+++ desk-mate-common/src/test/com/sunlands/deskmate/client/TestServiceTest.java
@@ -27,8 +27,8 @@ public class TestServiceTest {
     @Test
     public void test1() throws Exception {
 
-        List<Integer> list = new ArrayList<>();
-        list.add(1);
+        List<Long> list = new ArrayList<>();
+        list.add(1L);
         tzUserCenterService.findByIdIn(list);
     }
